requirements.txt
pandas
python-dotenv
serpapi
pendulum
datetime
google-search-results